@@ -2,7 +2,7 @@ import 'package:custom_widgets/widget/button/inset_button.dart';
 import 'package:flutter/material.dart';
 
 class ButtonsView extends StatefulWidget {
-  ButtonsView({Key? key}) : super(key: key);
+  const ButtonsView({super.key});
 
   @override
   State<ButtonsView> createState() => _ButtonsViewState();
@@ -14,25 +14,29 @@ class _ButtonsViewState extends State<ButtonsView> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Color(0xffe7ecef),
+      backgroundColor: const Color(0xffe7ecef),
       appBar: AppBar(
-        title: Text("Buttons"),
-        leading: IconButton(onPressed: () => Navigator.of(context).pop(), icon: Icon(Icons.arrow_back)),
+        title: const Text('Buttons'),
+        leading: IconButton(
+          onPressed: () => Navigator.of(context).pop(),
+          icon: const Icon(Icons.arrow_back),
+        ),
       ),
       body: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         crossAxisAlignment: CrossAxisAlignment.stretch,
         children: [
           Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 100),
-              child: InsetButton(
-                isInset: _isPressed,
-                onTap: () {
-                  setState(() {
-                    _isPressed = !_isPressed;
-                  });
-                },
-              )),
+            padding: const EdgeInsets.symmetric(horizontal: 100),
+            child: InsetButton(
+              isInset: _isPressed,
+              onTap: () {
+                setState(() {
+                  _isPressed = !_isPressed;
+                });
+              },
+            ),
+          ),
         ],
       ),
     );
