@@ -2,14 +2,14 @@ import 'package:custom_widgets/widget/custom_birth_picker.dart';
 import 'package:flutter/material.dart';
 
 class ListWheelView extends StatelessWidget {
-  const ListWheelView({Key? key}) : super(key: key);
+  const ListWheelView({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("ListWheelScroll"),
-        leading: IconButton(onPressed: () => Navigator.of(context).pop(), icon: Icon(Icons.arrow_back)),
+        title: const Text('ListWheelScroll'),
+        leading: IconButton(onPressed: () => Navigator.of(context).pop(), icon: const Icon(Icons.arrow_back)),
       ),
       body: Column(
         children: [
@@ -17,14 +17,16 @@ class ListWheelView extends StatelessWidget {
             flex: 5,
           ),
           Expanded(
-              flex: 2,
-              child: Center(
-                  child: FittedBox(
+            flex: 2,
+            child: Center(
+              child: FittedBox(
                 child: Text(
-                  "ListWheelScroll Date Picker",
-                  style: Theme.of(context).textTheme.bodyText1!.copyWith(fontSize: 24),
+                  'ListWheelScroll Date Picker',
+                  style: Theme.of(context).textTheme.bodyMedium!.copyWith(fontSize: 24),
                 ),
-              ))),
+              ),
+            ),
+          ),
           const Expanded(
             flex: 20,
             child: FractionallySizedBox(
